# docs/api/OPENAPI.yaml
openapi: 3.0.3
info:
  title: PEZHWAN Identity Server
  description: |
    Universal Identity & Access Management API. Reference host for the
    `@pezhwan/*` SDK stack. All endpoints use a `{ success, data, error }`
    envelope except OAuth/OIDC, which follow OAuth 2.1 / OIDC wire format.
  version: 0.1.0
servers:
  - url: http://localhost:4011
    description: Development reference server
security:
  - {}
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Password login (MFA-gated when enabled)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, example: demo@pezhwan.dev }
                phone: { type: string }
                password: { type: string }
              required: [password]
      responses:
        '200':
          description: Tokens (or MFA challenge when mfaRequired)
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/LoginTokens' }
                  - $ref: '#/components/schemas/MfaChallenge'
        '401':
          description: Invalid credentials / locked / disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /v1/auth/otp/login:
    post:
      tags: [Auth]
      summary: Passwordless OTP login (MFA-gated when enabled)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channel: { type: string, enum: [email, phone] }
                target: { type: string }
                code: { type: string }
              required: [channel, target, code]
      responses:
        '200':
          description: Tokens (or MFA challenge)
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/LoginTokens' }
                  - $ref: '#/components/schemas/MfaChallenge'
...
  /v1/mfa/setup:
    post:
      tags: [MFA]
      security:
        - Bearer: []
      summary: Begin TOTP setup (returns secret + otpauth URI + backup codes)
      responses:
        '200':
          description: Setup material
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string }
                  otpauthUri: { type: string }
                  backupCodes:
                    type: array
                    items: { type: string }
  /v1/mfa/enable:
    post:
      tags: [MFA]
      security:
        - Bearer: []
      summary: Confirm a TOTP code and enable MFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
              required: [code]
      responses:
        '200':
          description: Enabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
...
  /v1/mfa/disable:
    post:
      tags: [MFA]
      security:
        - Bearer: []
      summary: Disable MFA (requires current code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
              required: [code]
      responses:
        '200':
          description: Disabled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
  /v1/mfa/login:
    post:
      tags: [MFA]
      summary: Complete an MFA-challenged login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { type: string }
                code: { type: string }
              required: [userId, code]
      responses:
        '200':
          description: Tokens
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginTokens' }

  # ---------------------------------------------------------- Verification
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      description: Access token from /v1/auth/login|register|refresh (JWT RS256)
    ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: API key (pk_live_...) as `Bearer <key>`
    RoleAdmin:
      type: http
      scheme: bearer
      description: Requires the ADMIN role on the identity
  schemas:
    User:
      type: object
      properties:
        _id: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        emailVerified: { type: boolean }
        phoneVerified: { type: boolean }
        isActive: { type: boolean }
    LoginTokens:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        accessToken: { type: string }
        refreshToken: { type: string }
...
    MfaChallenge:
      type: object
      properties:
        mfaRequired: { type: boolean, enum: [true] }
        userId: { type: string }
...
    Ok:
      type: object
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          properties:
            message: { type: string }
    Error:
      type: object
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }